Sort indicators
The indicator tables in the PDF returned the indictors unpredictably, so
they were placed a single list and sorted them manually into the right
categories here. The process also threw up some problems with the regex
used in the script, which didn't match indictors spanning pages. So
the problems caused by this are fixed too.

The indicator index now needs fixing.