Add CAF 3.2 router
With the two field provider classes, the router is the other
framework-specific piece of code. It can handle versions of the
CAF which vary in content but not structure and small structural
changes could probably be handled by subclassing it.

However, if we wanted to accomomodate an assessment framework
which had a significantly different structure we would need
a replacement. An immediate next step is to create an ABC from which
this class can inherit to accomodate this.

The router can create a route with only organisation questions and
another with only system questions but this is currently unused.

The AppsConfig subclass is used to load the route.

Once the routes are loaded the router is no longer needed.