Enviornment fixes
Copy the CAF YAML into the container and specify the path to it in
the settings file.

Explicitly include the template and static paths for the webcaf app
in the settings file. This shouldn't be necessary but appears to be
so.

Add a Makefile to make it easier to run the dev server.