Add view factory
To create pages dynamically while not putting a lot conditional logic
in views, we can create a view class for each page. This also means
we can handle the routing as we would normally with a Django app,
by registering each of the view classes with urls.py.

At this stage all of our generated views are subclasses of FormView
but this may change and new factory functions will be needed.

It's a deliberate design choice that this module has know knowlege
of the CAF or CAFs.