Rename sections to outcomes
What were called "sections" in the YAML CAF are properly called outcomes.
The change has been made in the YAML 3.2 CAF, the fixtures CAF and the
CAF 3.2 -specific code.

We previously had an 'outcome' page and template for each
section/outcome but this has been renamed to 'confirmation' to avoid
confusion.