Simplify caf32_router
Before adding breadcrumb data to the views, remove redundant code from
the router. We now know we don't need system and org routes, so all
of the code supporting those can go. We also don't need the `all_route`
method to return anything in practice, so replace the tests which relied
on it returning a list of data about the pages added to the route with
tests which examine the thing we do need - additions to
webcaf.urls.urlpatterns.