Refactor FrameworkRouter
Add a placeholder ABC for FrameworkRouter and rename the latter to
reflect that it is a router for CAF 3.2.

Simplify the traversal of the framework and separate it from the
class's work, so that we can share the traversal code with a
prototype spreadsheet exporter. It's also in general more readable.