Add NCSC index to CAF YAML files
NCSC have applied an index to the indicators in their v4.0 changelog
document (https://www.ncsc.gov.uk/files/CAF%20v3.2%20Changelog.pdf).

The add-ncsc-index.py script adds this index to both the 3.2 and 4.0
YAML representations of the CAF.

The YAML files have been committed here with the new index added. For
now, this is in the form of an additional "ncsc_index" key under each
indicator but we may want to make it the overall indicator key.