Instantiate CAF40 router
Add a framework field to assessment to determine which version is being
used. For the time being this is hardcoded to 3.2.

Make the path to the framework document a property of the router class
and add a new router class for CAF4.0. Instantiate and execute both
routers on application load.

Add a new frameworks module from which the instantiated routers can be
accessed.