Fix urls for CAF versions
All the urls in the CAF route now have a path element for the CAF
version and each of the views, forms etc has a prefix in the name.

For a given assessment we can get the right path/name prefix from
the assessment model instance. Add this functionlity where needed.
The exception is when we create a new assessment instance, in which
cases we have to add this to the defaults passed to the template(s).

Start to separate out CAF-specific code from generic framework code
by adding abstract messages for 'main headings' to the router ABC
and adding concrete overriders in the CAF32 router. Make consequent
changes throughout the app. Remove the ConfigHelper class which
was doing this before.

One or two small fixes.