Separate out CAF-specific code
Create a new 'caf' package and move the vast majority of CAF-specific
code into it. The exceptions are some of the template tags which are
yet to be done.

The view factory is tightly bound to the CAF concepts (objectives etc)
so that goes into the new module. The form factory is more generic.

The purpose of this is to reduce the work involved in adding non-CAF
frameworks in the future. As things stand we do not have a clear
requirement for this but it is something we have been asked to
accomodate.