Change from character to word count in confirmation supporting comments
This change has resulted from round 1 of UR

The component has been updated to count words rather than characters with a limit
of 1500.

Previously there was no validation on this field so a custom validator class has been
added to validate word count aginst max allowed word count.

A further change has been made to the form element displayed, removing initial. This
is because otherwise on validation error the fields text is overwritten with the initial
value before more than 1500 words have been entered. Using form.supporting_comments
will use initial only if a current vaule in the form is not present. striptags is required
as otherwise some html is rendered in the field.

It also adds the error message partial to the confirmation template and validates the max
character limit against the maxlength set in the supporting_comments widget.

The id has also been updated from summary_outline to id_supporting_comments, so that
navigation from the error message takes a user back to the field with the error in it.