Add assessment review type page
This commit adds a new page for review type and 2 new views for create and edit views of the
new page.

It includes some additional logic as the selection of review type is dependent on the caf
profile selected, where if the selected profile is enhanced then the review type is set to
independent. Otherwise the user can select the review type from the provided options. The page
displays different content based on the caf profile selection too.

The Assessment model now has a review type field to capture the user selection of review type