Add progress bar to indicator page
This adds a progress bar, indicative of the % completed of the whole caf
self assessment, to the top of each indicator page

The progress box also gives user information around the number of principles
in the section and the pricinple the user is on i.e. 1 of 4.

Because no govuk component exists for the progress bar it implements some custom
css classes in a govuk style to achieve this and  match the design in the prototype
application.