Fix system selection bug
This fixes a bug where a system would reappear for selection after
an assessment for that system had been submitted. Meaning in theory
a new assessment could be started for an already completed assessment

Since the addition of the configuration model we can now reliably get
the current year so the query is simplfied to exclude any systems that
already have an assessment for the current year and this includes both
draf and submitted assessments.

Also add an additional feature test to cover this scenario