feat: Implement session timeout middleware
This commit adds code to handle the timeout and expiry of user
sessions.

`CafSessionTimeoutMiddleware` handles the logic for timing out a
user session, recording a last_activity in the session and then
checking whether a request from an authenticated user is greater
from last activity than timeout value specified in settings in the
`USER_IDLE_TIMEOUT` variable. Set to 90 minutes in this commit.

On session timeout the user is logged out of the app and redirected
to a session timeout page. There is a new simple view to handle this
redirect. The session timeout url path is also adding to the list of
paths not requiring a login.

It also adds an auto-refresh to the base template so that each page will
refresh shortly after 90mins of inactivty, redirecting to the session
timeout page.

The change addresses the UCD requirement for user sessions to be expired
after 90 mins, primarily related to enhanced security of the application